edit: refatoring components and structure paths
