+++ antd-demo/src/app/services/getAllBncc.ts
@@ -5,11 +5,14 @@ const getAllBncc = async () => {
     try{
         const response = await fetch(`${baseUrl}/api/options/bncc`);
         const data = await response.json();
+        
+        if(!response.ok){
+            throw new Error("Erro ao buscar bnccs!");
+        }
         return data;
     }
-    catch(e){
-        console.error(e);
-        return [];
+    catch(error: any){
+        throw new Error(error);
     };
 };
